I ate the leftovers and cleaned the plate.
I cleaned up the code quite a bit and removed my old main method. I now have an experimental mode for tweaking settings and a locked mode for competition. It turns out that the issue with my distance was that the cameras were not parallel as the pencil had fallen off.